Combine result classes for purposes of comprehension
